CTAA-1386 renaming to diagnosis keys
