+++ app/src/main/java/at/roteskreuz/stopcorona/screens/base/CoronaBaseActivity.kt
@@ -8,7 +8,7 @@ import androidx.annotation.LayoutRes
 import at.roteskreuz.stopcorona.R
 import at.roteskreuz.stopcorona.constants.isDebug
 import at.roteskreuz.stopcorona.model.entities.infection.message.MessageType
-import at.roteskreuz.stopcorona.screens.debug.exposure_notification_infection_messages.startDebugExposureNotificationsTracingKeysFragment
+import at.roteskreuz.stopcorona.screens.debug.diagnosis_keys.startDebugDiagnosisKeysFragment
 import at.roteskreuz.stopcorona.screens.debug.exposure_notifications.startDebugExposureNotificationsFragment
 import at.roteskreuz.stopcorona.skeleton.core.screens.base.activity.BaseActivity
 import org.koin.androidx.viewmodel.ext.android.getViewModel
@@ -86,8 +86,8 @@ open class CoronaBaseActivity(@LayoutRes layout: Int = R.layout.framelayout) : B
                 startDebugExposureNotificationsFragment()
                 true
             }
-            R.id.debugExposureNotificationsTrackingKeys -> {
-                startDebugExposureNotificationsTracingKeysFragment()
+            R.id.debugDiagnosisKeys -> {
+                startDebugDiagnosisKeysFragment()
                 true
             }
             else -> super.onOptionsItemSelected(item)
